backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
sqlmodel
alembic
httpx
openai
mcp
tenacity
redis
pytest
pytest-asyncio
pytz
requests
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
